django_cleanup/testapp/requirements.txt
django>1.7,<1.8
sorl-thumbnail
easy-thumbnails
pillow
flexmock
pytest
pytest-django
pytest-pythonpath
pytest-cov
